core_py/requirements.txt
aiofiles==24.1.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
cryptography==45.0.6
distlib==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
identify==2.6.13
idna==3.10
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
nodeenv==1.9.1
numpy==2.2.6
oauthlib==3.3.1
ollama==0.5.3
openai==1.98.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.47.2
tabulate==0.9.0
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
wsproto==1.2.0
